+++ plugins/source/aws/resources/services/route53/health_checks.go
@@ -12,7 +12,7 @@ func HealthChecks() *schema.Table {
 		Description: `https://docs.aws.amazon.com/Route53/latest/APIReference/API_HealthCheck.html`,
 		Resolver:    fetchRoute53HealthChecks,
 		Transform:   transformers.TransformWithStruct(&Route53HealthCheckWrapper{}, transformers.WithUnwrapStructFields("HealthCheck")),
-		Multiplex:   client.AccountMultiplex,
+		Multiplex:   client.ServiceAccountRegionMultiplexer("route53"),
 		Columns: []schema.Column{
 			client.DefaultAccountIDColumn(false),
 			{