+++ plugins/source/postgresql/resources/plugin/tables.go
@@ -1,0 +1,13 @@
+package plugin
+
+import (
+	"context"
+
+	"github.com/cloudquery/cloudquery/plugins/source/postgresql/client"
+	"github.com/cloudquery/plugin-sdk/schema"
+)
+
+func getDynamicTables(ctx context.Context, c schema.ClientMeta) (schema.Tables, error) {
+	cl := c.(*client.Client)
+	return cl.Tables, nil
+}