+++ plugins/destination/s3/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [2.2.1](https://github.com/cloudquery/cloudquery/compare/plugins-destination-s3-v2.2.0...plugins-destination-s3-v2.2.1) (2023-02-14)
+
+
+### Bug Fixes
+
+* **deps:** Update module github.com/cloudquery/filetypes to v1.4.1 ([#8024](https://github.com/cloudquery/cloudquery/issues/8024)) ([380476f](https://github.com/cloudquery/cloudquery/commit/380476fc11cb5cab576f320baa12e215bb148f86))
+* **deps:** Update module github.com/cloudquery/plugin-sdk to v1.37.1 ([#8008](https://github.com/cloudquery/cloudquery/issues/8008)) ([c47aac0](https://github.com/cloudquery/cloudquery/commit/c47aac0b5e3190a04299713651b97e360043911f))
+
+## [2.2.0](https://github.com/cloudquery/cloudquery/compare/plugins-destination-s3-v2.1.0...plugins-destination-s3-v2.2.0) (2023-02-14)
+
+
+### Features
+
+* Add support for date placeholders in S3 plugin ([#7981](https://github.com/cloudquery/cloudquery/issues/7981)) ([7d3c25f](https://github.com/cloudquery/cloudquery/commit/7d3c25fccd829595d1725097a8c6e034353b8ec5))
+
+
+### Bug Fixes
+
+* **deps:** Update module github.com/cloudquery/plugin-sdk to v1.36.0 ([#7809](https://github.com/cloudquery/cloudquery/issues/7809)) ([c85a9cb](https://github.com/cloudquery/cloudquery/commit/c85a9cb697477520e94a1fd260c56b89da62fc87))
+* **deps:** Update module github.com/cloudquery/plugin-sdk to v1.37.0 ([#7933](https://github.com/cloudquery/cloudquery/issues/7933)) ([dc9cffb](https://github.com/cloudquery/cloudquery/commit/dc9cffbf37bbc6fae73a20bf47e6bbf17e74d1f9))
+
 ## [2.1.0](https://github.com/cloudquery/cloudquery/compare/plugins-destination-s3-v2.0.0...plugins-destination-s3-v2.1.0) (2023-02-07)